Fix index overflow

+++ cupy/core/core.pyx
@@ -4326,9 +4326,9 @@ def _partition_kernel(dtype):
     source = string.Template('''
     template<typename T>
     __device__ void bitonic_sort_step(CArray<T, 1> a,
-            int x, int y, int i, int s, int w) {
-        for (int j = i; j < (y - x) / 2; j += 32) {
-            int n = j + (j & -w);
+            ptrdiff_t x, ptrdiff_t y, int i, ptrdiff_t s, ptrdiff_t w) {
+        for (ptrdiff_t j = i; j < (y - x) / 2; j += 32) {
+            ptrdiff_t n = j + (j & -w);
             T v = a[n + x], u = a[n + w + x];
             if (n & s ? v < u : v > u) {
                 a[n + x] = u;
@@ -4339,9 +4339,10 @@ def _partition_kernel(dtype):
 
     // Sort a[x:y].
     template<typename T>
-    __device__ void bitonic_sort(CArray<T, 1> a, int x, int y, int i) {
-        for (int s = 2; s <= y - x; s *= 2) {
-            for (int w = s / 2; w >= 1; w /= 2) {
+    __device__ void bitonic_sort(
+            CArray<T, 1> a, ptrdiff_t x, ptrdiff_t y, int i) {
+        for (ptrdiff_t s = 2; s <= y - x; s *= 2) {
+            for (ptrdiff_t w = s / 2; w >= 1; w /= 2) {
                 bitonic_sort_step<T>(a, x, y, i, s, w);
             }
         }
@@ -4349,7 +4350,8 @@ def _partition_kernel(dtype):
 
     // Merge first k elements and the next 32 times t elements.
     template<typename T>
-    __device__ void merge(CArray<T, 1> a, int k, int i, int x, int z, int u) {
+    __device__ void merge(
+            CArray<T, 1> a, int k, int i, ptrdiff_t x, ptrdiff_t z, int u) {
         for (int s = i; s < u; s += 32) {
             if (a[x + k - s - 1] > a[z + s]) {
                 T tmp = a[x + k - s - 1];
@@ -4370,17 +4372,18 @@ def _partition_kernel(dtype):
     // the warp size. The first k elements are always sorted and the next 32
     // times t elements stored values that have possibilities to be selected.
     __global__ void ${name}(
-            CArray<${dtype}, 1> a, int k, int n, int t, int sz) {
+            CArray<${dtype}, 1> a, int k, ptrdiff_t n, int t, ptrdiff_t sz) {
 
         // This thread handles a[z:m].
-        long long i = blockIdx.x * blockDim.x + threadIdx.x;
-        int z = i / 32 * n / sz;
-        int m = (i / 32 + 1) * n / sz;
+        ptrdiff_t i = static_cast<ptrdiff_t>(blockIdx.x) * blockDim.x
+            + threadIdx.x;
+        ptrdiff_t z = i / 32 * n / sz;
+        ptrdiff_t m = (i / 32 + 1) * n / sz;
         int id = i % 32;
         int x = 0;
 
         bitonic_sort<${dtype}>(a, z, k + z, id);
-        int j;
+        ptrdiff_t j;
         for (j = k + id + z; j < m - (m - z) % 32; j += 32) {
             if (a[j] < a[k - 1 + z]) {
                 ${dtype} tmp = a[k + 32 * x + id + z];
@@ -4414,10 +4417,11 @@ def _partition_kernel(dtype):
     }
 
     __global__ void ${merge_kernel}(
-            CArray<${dtype}, 1> a, int k, int n, int sz, int s) {
-        long long i = blockIdx.x * blockDim.x + threadIdx.x;
-        int z = i / 32 * 2 * s * n / sz;
-        int m = (i / 32 * 2 + 1) * s * n / sz;
+            CArray<${dtype}, 1> a, int k, ptrdiff_t n, int sz, int s) {
+        ptrdiff_t i = static_cast<ptrdiff_t>(blockIdx.x) * blockDim.x
+            + threadIdx.x;
+        ptrdiff_t z = i / 32 * 2 * s * n / sz;
+        ptrdiff_t m = (i / 32 * 2 + 1) * s * n / sz;
         int id = i % 32;
         merge<${dtype}>(a, k, id, z, m, k);
     }
